+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * fix: increase socket timeout to 10 minutes
 * fix: allow sync forcing updates using UI
 * fix: only show sync button if there are no songs playing or queued
+* feature: display date created in song list
 
 ## 1.1-2019.10.29-SNAPSHOT
 * retrieve audio file header information (like file length)